j
j

+++ script.js
@@ -3,7 +3,7 @@ let registroNotas = [];
 
 function notasEstudiantes() {
     for (let i = 0; i < estudiantes.length; i++) {
-        let notaAlumno = parseInt(prompt(`Ingrese la nota final de ${estudiantes[i]}:`));
+        let notaAlumno = parseFloat(prompt(`Ingrese la nota final de ${estudiantes[i]}:`));
         
         if (notaAlumno >= 4 && notaAlumno < 8) {
             alert(`${estudiantes[i]} aprobada/o`);
